Zombies can now path find. Works suprisingly well.

+++ ZombieSquad/Vec2.h
@@ -35,7 +35,7 @@ public: // methods
 	// return the vector length
 	T Length() const
 	{
-		T ret = sqrt(SqLength());
+		T ret = sqrtf(SqLength());
 		return ret;
 	}
